ml_2/requirements.txt
# ATLAS ML Pipeline - Dependencies
# ================================

# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0

# OCR
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# Vision-Language Model (LLaVA)
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0  # For quantization

# Local LLM → Now using Gemini API
google-generativeai>=0.8.0

# Computer Vision
opencv-python>=4.8.0
numpy>=1.24.0

# Screen Capture & Control
mss>=9.0.0
pyautogui>=0.9.54
pillow>=10.0.0

# Agent Orchestration
langgraph>=0.0.20

# Windows-specific (for window management)
pywin32>=306
uiautomation>=2.0.18

# Utilities
pydantic>=2.0.0  # For config/schema validation
python-dotenv>=1.0.0  # Environment config
loguru>=0.7.0  # Logging
